add apache license for all ts file
