+++ control-panel/NOTICE
@@ -1,0 +1,5 @@
+Apache ServiceComb Service Center
+Copyright 2017-2019 The Apache Software Foundation
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
