--- control-panel/cp-frontend/src/app/author/author.component.spec.ts
@@ -1,25 +1,0 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-
-import {AuthorComponent} from './author.component';
-
-describe('AuthorComponent', () => {
-  let component: AuthorComponent;
-  let fixture: ComponentFixture<AuthorComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [AuthorComponent]
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(AuthorComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
