--- control-panel/cp-frontend/src/app/home/home.component.html
@@ -1,40 +1,0 @@
-<div>
-  <mat-card style="margin: 10vh auto; width: 400px;">
-    <mat-card-header>
-      <mat-card-title>Below text version is HTTP GET from the API call</mat-card-title>
-      <mat-card-subtitle>API url: http://localhost:3000/api/version</mat-card-subtitle>
-    </mat-card-header>
-    <mat-card-content>
-      <p>
-        The application version is {{ version }}
-      </p>
-    </mat-card-content>
-  </mat-card>
-
-  <div class="micro-service-scope"  fxLayout="row" fxLayoutAlign="space-around">
-    <mat-card style="width:300px;" *ngFor="let ms of services">
-      <mat-card-header>
-        <mat-card-title>ServiceName: {{ ms.serviceName }}</mat-card-title>
-        <mat-card-subtitle>Version: {{ ms.version }}</mat-card-subtitle>
-      </mat-card-header>
-      <mat-card-content>
-        <p>
-          ServiceId: {{ms.serviceId}}
-        </p>
-      </mat-card-content>
-    </mat-card>
-  </div>
-
-  <mat-card style="margin: 10vh auto; width: 1000px;">
-    <mat-card-header>
-      <mat-card-title>Below text version is pushed by server through websocket</mat-card-title>
-      <mat-card-subtitle>websocket url: http://localhost:3000/websocket</mat-card-subtitle>
-    </mat-card-header>
-    <mat-card-content>
-      <p><b>events from server:</b></p>
-      <p *ngFor="let event of events">
-        {{event.action}} -- {{event.key.appId}} -- {{event.key.serviceName}}
-      </p>
-    </mat-card-content>
-  </mat-card>
-</div>
