--- control-panel/cp-frontend/src/app/model/micro-service.model.ts
@@ -1,6 +1,0 @@
-export class MicroServiceModel {
-  serviceId: string;
-  appId: string;
-  serviceName: string;
-  version: string;
-}
