--- control-panel/cp-frontend/src/app/model/server-event.model.ts
@@ -1,4 +1,0 @@
-export class ServerEventModel {
-  action: string;
-  key: {tenant: string, appId: string, serviceName: string, version: string};
-}
