--- control-panel/cp-frontend/src/app/model/version.model.ts
@@ -1,4 +1,0 @@
-export class VersionModel {
-  name: string;
-  tag: string;
-}
