--- control-panel/cp-frontend/src/app/service/websocket.service.spec.ts
@@ -1,16 +1,0 @@
-import {TestBed} from '@angular/core/testing';
-
-import {WebsocketService} from './websocket.service';
-
-describe('WebsocketService', () => {
-  let service: WebsocketService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(WebsocketService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
